requirements.txt
flask==3.0.3
pytest==8.0.0
python-dotenv==1.0.0